wingz/requirements.txt
asgiref==3.8.1
astroid==3.3.8
dill==0.3.9
Django==5.1.6
django-phonenumber-field==8.0.0
django-rest-framework==0.1.0
djangorestframework==3.15.2
iniconfig==2.0.0
isort==6.0.1
mccabe==0.7.0
packaging==24.2
phonenumberslite==8.13.55
platformdirs==4.3.6
pluggy==1.5.0
pylint==3.3.4
pylint-django==2.6.1
pylint-plugin-utils==0.8.2
pytest==8.3.4
pytest-django==4.10.0
python-decouple==3.8
sqlparse==0.5.3
tomlkit==0.13.2
